% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kappa.R
\name{cut.Kappa}
\alias{cut.Kappa}
\title{cut.Kappa}
\usage{
\method{cut}{Kappa}(x, ...)
}
\arguments{
\item{x}{a \link[vcd]{Kappa} object}

\item{...}{trivial parameters of function \link[base]{cut.default}}
}
\description{
cut.Kappa
}
\note{
Hard-coded cut-off values from \url{https://en.wikipedia.org/wiki/Cohen\%27s_kappa#Interpreting_magnitude}.
}
\keyword{internal}
