% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kappa.R
\name{Kappa2}
\alias{Kappa2}
\title{Kappa with input confusion matrix}
\usage{
Kappa2(x, ...)
}
\arguments{
\item{x}{see function \link[vcd]{Kappa}}

\item{...}{additional parameters of function \link[vcd]{Kappa}}
}
\description{
Kappa with input confusion matrix
}
\details{
Function \link[vcd]{Kappa} does not keep any info of the input (eh..)
}
\examples{
data('JobSatisfaction', package = 'vcd')
(job1 = xtabs(Freq ~ supervisor + own, data = JobSatisfaction))
(job2 = xtabs(Freq ~ management + own, data = JobSatisfaction))

list(
  'supervisor vs. own' = job1 |> vcd::Kappa(),
  'management vs. own' = job2 |> vcd::Kappa()
) |> rmd.tzh::render_(file = 'Kappa')

list(
  'supervisor vs. own' = job1 |> Kappa2(),
  'management vs. own' = job2 |> Kappa2()
) |> rmd.tzh::render_(file = 'Kappa2')

list(
  'sec 1' = list(
    'abc',
    job1 |> Kappa2()
  )
) |> rmd.tzh::render_(file = 'Kappa2_v2')

}
\keyword{internal}
