% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kappa.R
\name{md_.Kappa}
\alias{md_.Kappa}
\title{R Markdown Lines for \link[vcd]{Kappa} or \link{Kappa2}}
\usage{
\method{md_}{Kappa}(x, xnm, ...)
}
\arguments{
\item{x}{\link[vcd]{Kappa} or \link{Kappa2} object}

\item{xnm}{..}

\item{...}{..}
}
\description{
R Markdown Lines for \link[vcd]{Kappa} or \link{Kappa2}
}
\keyword{internal}
