% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kappa.R
\name{S3_Kappa}
\alias{S3_Kappa}
\alias{.pval.Kappa}
\alias{endpoint.Kappa}
\title{Additional S3 Methods for \link[vcd]{Kappa}}
\usage{
\method{.pval}{Kappa}(x)

\method{endpoint}{Kappa}(x)
}
\arguments{
\item{x}{a \link[vcd]{Kappa} object}
}
\description{
Additional S3 Methods for \link[vcd]{Kappa}
}
\keyword{internal}
