% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib.R
\name{vcd_bib}
\alias{vcd_bib}
\alias{.cohen60}
\title{bibs in vcd package}
\usage{
.cohen60(key = "Cohen60", ...)
}
\arguments{
\item{key, ...}{additional parameters of function \link[utils]{bibentry}}
}
\description{
bibs in vcd package
}
\keyword{internal}
