% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kappa.R
\name{cut.Kappa}
\alias{cut.Kappa}
\title{cut.Kappa}
\usage{
\method{cut}{Kappa}(
  x,
  breaks = c(-Inf, 0, 0.2, 0.4, 0.6, 0.8, 1),
  labels = c("no", "slight", "fair", "moderate", "substantial", "almost perfect"),
  right = TRUE,
  include.lowest = TRUE,
  ...
)
}
\arguments{
\item{x}{a \link[vcd]{Kappa} object}

\item{breaks, labels, right, include.lowest, ...}{parameters of function \link[base]{cut.default}}
}
\description{
cut.Kappa
}
\examples{
library(vcd)
(job1 = xtabs(Freq ~ supervisor + own, data = JobSatisfaction))
(job2 = xtabs(Freq ~ management + own, data = JobSatisfaction))
job1 |> Kappa() |> cut()
job2 |> Kappa() |> cut()
}
\keyword{internal}
